Add SellShop support

+++ ShopIndex/Models/ShopSync/Location.cs
@@ -2,7 +2,7 @@
 
 public class Location
 {
-    public int[] Coordinates { get; set; } = new int[3];
+    public float[] Coordinates { get; set; } = new float[3];
     public string? Description { get; set; }
     public string Dimension { get; set; } = "overworld";
 }
